--- a/crates/iddqd/src/support/map_hash.rs
+++ b/crates/iddqd/src/support/map_hash.rs
@@ -19,11 +19,13 @@ impl<S> fmt::Debug for MapHash<S> {
 }
 
 impl<S: BuildHasher> MapHash<S> {
+    #[cfg(feature = "std")]
     #[inline]
     pub(crate) fn build_hasher(&self) -> S::Hasher {
         self.state.build_hasher()
     }
 
+    #[cfg(feature = "std")]
     #[inline]
     pub(crate) fn hash(&self) -> u64 {
         self.hash