--- a/scripts/miri-ignore-leaks.sh
+++ b/scripts/miri-ignore-leaks.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+
+# A nextest wrapper script that instructs Miri to ignore leaks.
+
+MIRIFLAGS="${MIRIFLAGS} -Zmiri-ignore-leaks" "$@"