This allows `reborrow` to work against borrowed data (though based on https://github.com/rust-lang/rust/issues/92985 the new Polonius implementation would also solve this issue), and means that the type definition doesn't have to carry around the `Hash` restriction.

As one might expect from dynamic dispatch, this does slow down the benchmark:

```
before:
id_ord_map_ref_mut_simple
                        time:   [72.419 ns 72.574 ns 72.739 ns]

after:
id_ord_map_ref_mut_simple
                        time:   [79.710 ns 79.984 ns 80.263 ns]
                        change: [+8.2749% +9.6421% +11.091%] (p = 0.00 < 0.05)
                        Performance has regressed.
```

So it's a judgment call about whether we should go in this direction.