@sunshowers 2025-08-12
Been poking around at this, and:

* The `'a` in `RefMut<'a, T>` is currently covariant -- which means it can be shortened to a `'b`, even without reborrowing
* storing a `fn(&foldhash::fast::RandomState, T::Key<'a>) -> u64` directly will make `'a` invariant, which is inconvenient and also a breaking change.
* in order to make `'a` covariant again, we'd have to store something like a `*const ()` instead and cast it to the appropriate function pointer as required.
* but the covariance can lead to what I think is UB, in case `Hash` is only implemented for `T::Key<'static>`.

Here's an example:

```rust
mod static_breakage {
    use std::hash::Hash;

    use super::*;

    struct Item {
        id: String,
    }

    impl IdOrdItem for Item {
        type Key<'a> = Id<'a>;

        fn key(&self) -> Self::Key<'_> {
            Id(&self.id)
        }

        id_upcast!();
    }

    #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord)]
    struct Id<'a>(&'a str);

    impl Hash for Id<'static> {
        fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
            // XXX this gets called for a shorter Id<'a> while dropping the
            // Foo.
            self.0.hash(state);
        }
    }

    #[test]
    fn static_breakage() {
        let mut map = IdOrdMap::new();
        map.insert_overwrite(Item { id: "test".to_string() });

        let map: &'static mut IdOrdMap<Item> = Box::leak(Box::new(map));
        let item = map.get_mut(&Id("test")).unwrap();

        let s = "test".to_owned();
        let foo = Foo { item, s: &s };
    }

    struct Foo<'a> {
        item: id_ord_map::RefMut<'a, Item>,
        s: &'a str,
    }
}
```

Now it is only possible to call `get_mut` on a map available for the `'static` lifetime, due to the `T::Key<'a>: Hash` requirement. So the underlying string is always `'static` regardless of the lifetime at the time it's being dropped. But the fact that this is possible scares me, and makes me think we should abandon this approach.